リファクタリング: `/generate_image` と `/send_today_plan` エンドポイントを `routes/generate.py` と `routes/suggest.py` に移動し、`app.py` から不要なインポートを削除。テスト用の提案データを `data/test_outfit_proposals.json` に追加。
