+++ data/test_outfit_proposals.json
@@ -1,0 +1,82 @@
+{
+  "proposals": [
+    {
+      "pattern": 1,
+      "items": {
+        "tops": {
+          "id": 6,
+          "item_name": "白シャツ",
+          "image_path": "/static/images/1/clothes/t003.jpg",
+          "taste": [
+            "きれいめ",
+            "カジュアル",
+            "フォーマル"
+          ]
+        },
+        "bottoms": {
+          "id": 13,
+          "item_name": "黒スラックス",
+          "image_path": "/static/images/1/clothes/b002.jpg",
+          "taste": [
+            "きれいめ",
+            "フォーマル"
+          ]
+        },
+        "outer": {
+          "id": 18,
+          "item_name": "ネイビーテーラードジャケット",
+          "image_path": "/static/images/1/clothes/o001.jpg",
+          "taste": [
+            "きれいめ",
+            "フォーマル"
+          ]
+        }
+      },
+      "item_ids": [
+        6,
+        13,
+        18
+      ],
+      "reason": "22℃でディナー、降水確率90%という条件に対し、室内での食事をメインに想定した、 きちんと感のあるきれいめコーディネートです。ホ ワイトの長袖シャツとブラックスラックスは、気温 に合っており、ディナーに相応しい上品さがありま す。ブラックのテーラードジャケットを羽織ること で、フォーマルな場にも適応しつつ、ポリエステル 素材は小雨程度なら対応しやすいでしょう。全体を モノトーンでまとめ、洗練された印象を与えます。"
+    },
+    {
+      "pattern": 2,
+      "items": {
+        "tops": {
+          "id": 10,
+          "item_name": "サックスブルーシャツ",
+          "image_path": "/static/images/1/clothes/t007.jpg",
+          "taste": [
+            "きれいめ",
+            "カジュアル"
+          ]
+        },
+        "bottoms": {
+          "id": 16,
+          "item_name": "黒デニムパンツ",
+          "image_path": "/static/images/1/clothes/b005.jpg",
+          "taste": [
+            "カジュアル",
+            "きれいめ"
+          ]
+        },
+        "outer": {
+          "id": 20,
+          "item_name": "ベージュトレンチコート",
+          "image_path": "/static/images/1/clothes/o003.jpg",
+          "taste": [
+            "きれいめ",
+            "トラッド"
+          ]
+        }
+      },
+      "item_ids": [
+        10,
+        16,
+        20
+      ],
+      "reason": "22℃で雨のディナーに対応する、 きれいめカジュアルなコーディネートです。ライト ブルーのシャツとブラックスリムデニムで、清潔感 とスマートさを演出します。アウターにはベージュ のトレンチコートを選び、ディナーの場にふさわし い上品さと季節感をプラス。綿素材ですが、22℃という気温には適しており、降水確率90%に対しては傘を併用することで対応し、屋内に入れば脱いで快適に 過ごすことを想定しています。"
+    }
+  ]
+}
+
